requirements.txt
gradio
torch
lightning
pillow
plotly
einops
torchvision
transformers